Handle `nostrUris` when parsing Uris
Move ConversionUtils to `core:utils`;
Create Constants file;

Create NostrUriUtils in `core:utils` and move nostrUri logic there
    Not all logic could be moved. Some of it depends on Nip19. That was
    left in `NostrResources` and should be reconsidered.
