+++ core/utils/build.gradle.kts
@@ -71,6 +71,12 @@ kotlin {
                 implementation(libs.kotlinx.datetime)
                 implementation(libs.kotlinx.serialization.json)
                 implementation(libs.ktor.http)
+
+                implementation(libs.napier)
+
+                // Cryptography
+                implementation(libs.bitcoin.kmp)
+                implementation(libs.secp256k1.kmp)
             }
         }
