+++ core/utils/src/commonMain/kotlin/net/primal/core/utils/Constants.kt
@@ -1,0 +1,10 @@
+package net.primal.core.utils
+
+const val NOSTR = "nostr:"
+const val NPUB = "npub1"
+const val NSEC = "nsec1"
+const val NEVENT = "nevent1"
+const val NADDR = "naddr1"
+const val NOTE = "note1"
+const val NRELAY = "nrelay1"
+const val NPROFILE = "nprofile1"
