+++ domain/nostr/src/commonMain/kotlin/net/primal/domain/nostr/Nip19TLV.kt
@@ -7,8 +7,8 @@ import io.ktor.utils.io.core.toByteArray
 import io.ktor.utils.io.core.writeFully
 import kotlinx.io.readByteArray
 import kotlinx.io.readString
-import net.primal.domain.nostr.cryptography.bechToBytesOrThrow
-import net.primal.domain.nostr.cryptography.toHex
+import net.primal.core.utils.bechToBytesOrThrow
+import net.primal.core.utils.toHex
 
 object Nip19TLV {
     enum class Type(val id: Byte) {
