+++ domain/nostr/src/commonMain/kotlin/net/primal/domain/nostr/utils/StringUtils.kt
@@ -2,7 +2,7 @@ package net.primal.domain.nostr.utils
 
 import fr.acinq.secp256k1.Hex
 import net.primal.core.utils.ellipsizeMiddle
-import net.primal.domain.nostr.cryptography.toNpub
+import net.primal.core.utils.toNpub
 
 fun String.asEllipsizedNpub(): String = Hex.decode(this).toNpub().ellipsizeMiddle(size = 8)
